Add feature flag apis.
I'm going to start work on workstreams. I don't want that to get in the way of launching the UI. That is, I want to launch as soon as the current integration work is done.

Putting workstream related work behind a feature flag will allow us to work on it, but still ship even when it's half implemented.