Breaking Change
Turn.LastCommitHash and Turn.BaselineCommitHash are removed and
replaced with Turn.CommitInfo, which is a map from org/repo to a
CommitInfo struct (which has LastCommitHash and BaseLineCommitHash
fields).

This enables turns in envs with more than one repo.

I have a follow up to the service for this change as well.

We will need to:

1. Checkin this change.
2. Update dependencies in the service, and check in changes to UpdateTurn to accomdate this.
3. Update dependencies in agent-wrapper, and change it's code for reading / updating these fields.

Then re-create the DB (to remove old turns that no have incompatible json data stored in them).
Deploy the update service.
Update the :dev tag for the agent wrapper to point at the latest agent version.

This pr needs to go in first.