Update api spec for Workstreams
We need to support a list of short-names for workstreams, with one of them being marked as the default.
This allows us to refer to tasks using a workstream short name and a task number. For example, a workstream named
"Organizations Support" might have a short name of "ORG", so that we can refer to tasks as "ORG-1", "ORG-2", etc.

We are doing this, because once our tool is used for "sprint planning" purposes, people will need to refer to tasks
externally, in slack, spred sheets, etc. Using a name like "ORG-3924" is much easier than
"ABBA5F16-27A2-46A8-9F8F-A1D002B958C5".

Short names must be unique across a tenant, and can contain only uppercase letters, numbers, and hyphens. They must
start with a letter, and be at least 2 characters long.

The UI will use AI to generate a short name given the initial title of the workstream and the short names of existing
workstreams.

We will allow the user to change the short name of an existing workstream if they don't like the one we picked. To
preserve compatability, they may also add a new name and mark it as the default.
This will allow old short names to continue to reference tasks. For example, it's possible for OLD-1234 and NEW-1234
to refer to the same task.

If a user selects a short name that matches that of an existing workstream, the api will return an error.
In the UI, we will display an error indicating the name is already in use. If the conflicting workstream is archived,
we will give them the option to "remove" the short name from the archived workstream. If we remove the last short name
from an archived workstream, we will add a default short name of "ARCHIVED-<first 5 chars of workstream id>" to that
workstream, so that we can still display short names for tasks when the user views archived workstreams. If the
"ARCHIVED-<first 5 chars of workstream id>" short names conflicts, we will append a number to it, like
"ARCHIVED-<first 5 chars>-2".

Also, we need to add a new CompletedAt field to the Turn object. This will be set when a turn is marked as completed.
We update the "UpdatedAt" field for turns when we update logs, and we can update logs after a turn has completed (there
can be buffered log lines we are in the process of up-loading). We update the version / updated time on the turn when
uploading logs so that we can ensure logs are properly sequenced (if there are parallel callers, one will fail with a
conflict). So, we can't rely on "UpdatedAt" to know when a turn is completed, so we need a separate CompletedAt field.

We present AI starting a turn and completing a turn as separate "Activities" in a work stream task's "activity" list,
just like we would for comments on tasks assigned to humans.

The purpose of this task is to update the API spec to reflect the changes above. This requires us to:

1. Add DefaultShortName and TaskCounter fields to the Workstream object.
2. Update the docs for CreateWorkstream, ListWorkstreams, GetWorkstream and UpdateWorkstream to reflect the new fields.
   (TaskCounter is read-only and managed by the service.)
3. Add methods for managing short names attached to workstreams (list, add, delete). Note there is no get api.
4. Add a TaskNumber field to the Task object. Its value is readonly via tha api, and will be updated by the service.
   We need to update the docs for the task related apis to reflect this. Note that this field will be nil for "normal"
   tasks (not in workstreams). A task can will re-numbered if it is moved
   from one workstream to another. If a task is moved out of a workstream, then moved back, it's number will be
   different from what it was before.
5. Add an CompletedAt field to the Turn object, and update the turn apis accordingly.
6. Add a MoveTask api that can move a task from one workstream to another. This will update the TaskNumber field
   on the task to the next number in the destination workstream. It will also increment the TaskCounter on the destination
   workstream.
7. Add a MoveShortName api, that can atomically:
   a. Remove a short name from one workstream
   b. Add that short name to another workstream, and optionally mark it as the default.
   c. Add a new short name to the source workstream.

Note that deleting short names from a workstream is a "hard" delete, not a "soft" delete.