Add docs for new APIs to support runners and github connections.
Also delete apis from the docs that will be deprecated. The deprecated
apis are still present in the sdk and the sevice implemetnation, and
won't be deleted until the replacements are implemented and the ui is
modified to switch to them.

Finally, update docs to reflect a default maxResults value of 10. A value of 500 is too hight to use as the default.